p_setup: widen linedef sidedef/vertex indices to 32 bits for large UDMF maps
ZDoom UDMF maps can declare far more than the 65535 sidedefs (and 65535
vertices) that a 16-bit index can address -- MyHouse.pk3's single map has
~301k sidedefs.  line_t.sidenum was unsigned short and the UDMF linedef
loader truncated sidefront/sideback (and the v1/v2 vertex indices) with
(unsigned short) casts, so every linedef whose sidedef index exceeded
65535 silently referenced sidenum mod 65536 -- the wrong sidedef, hence
the wrong sector.  At the player start this put the subsector in the
wrong sector (wrong floor height, spawned stuck) and drew the wrong
sectors entirely.

Widen line_t.sidenum to int, make NO_INDEX -1 (65535 is now a valid
index), widen P_FinalizeLineDef's vertex parameters to int, drop the
truncating casts in the UDMF loader, and translate the binary formats'
on-disk 0xffff no-sidedef marker to NO_INDEX.  Doom/Heretic/Hexen binary
maps are unaffected (their indices fit in 16 bits); doom2 first-frame
render is byte-identical.

+++ src/doomdata.h
@@ -109,7 +109,11 @@ typedef struct {
   unsigned short sidenum[2];
 } PACKEDATTR hexen_maplinedef_t;
 
-#define NO_INDEX ((unsigned short)-1)
+/* -1 rather than 0xffff: line_t.sidenum is now a 32-bit int (UDMF maps can
+ * have >65535 sidedefs, so 65535 is a valid index and cannot be the sentinel).
+ * The on-disk binary formats still use 0xffff for "no sidedef"; the loaders
+ * translate that to NO_INDEX. */
+#define NO_INDEX (-1)
 
 //
 // LineDef attributes.
